Switch blocking writes to the subprocess to run on a dedicated executor.
The dedicated executor will be auto created by the manager, or can be
specified at manager creation time if the caller wants to control the
lifecycle. Also adds regression test for large input/output to the
terminal and switches basic_test.py's fileformat to unix from dos so
that it matches with the rest of the source code.